fixed the issue -> now streaming on grafana

+++ db/schema.sql
@@ -1,8 +1,8 @@
 CREATE TABLE IF NOT EXISTS traffic_metrics(
     window_start TIMESTAMP,
-    requests_per_min INTEGER,
+    requests_per_min BIGINT,
     error_rate DOUBLE PRECISION,
     avg_response_time DOUBLE PRECISION,
-    unique_users INTEGER,
+    unique_users BIGINT,
     is_anomaly BOOLEAN
 )
